Update and modify files for the project - Added new files:   - Ethernaut/05-Token_DONE/.gitignore   - Ethernaut/05-Token_DONE/ape-config.yaml   - Ethernaut/05-Token_DONE/contracts/Token.sol   - Ethernaut/05-Token_DONE/poetry.lock   - Ethernaut/05-Token_DONE/pyproject.toml   - Ethernaut/05-Token_DONE/scripts/attack.py   - Ethernaut/06-Delegation_DONE/.gitignore   - Ethernaut/06-Delegation_DONE/ape-config.yaml   - Ethernaut/06-Delegation_DONE/contracts/Delegation.sol   - Ethernaut/06-Delegation_DONE/poetry.lock   - Ethernaut/06-Delegation_DONE/pyproject.toml   - Ethernaut/06-Delegation_DONE/scripts/attack.py - Updated files:   - Ethernaut/Coin_Flip_DONE/contracts/CoinFlip.sol   - Ethernaut/Coin_Flip_DONE/poetry.lock   - Ethernaut/Coin_Flip_DONE/pyproject.toml   - Ethernaut/Delegation_DONE/contracts/Delegation.sol   - Ethernaut/Fallback_DONE/contracts/Fallback.sol   - Ethernaut/Fallout_DONE/contracts/Fallout.sol   - Ethernaut/Telephone_DONE/contracts/Attack.sol   - Ethernaut/Telephone_DONE/contracts/Telephone.sol

+++ CTFs/Ethernaut/Fallout_DONE/contracts/Fallout.sol
@@ -5,7 +5,8 @@ import "@openzeppelin/math/SafeMath.sol";
 
 contract Fallout {
     using SafeMath for uint256;
-    mapping(address => uint) allocations;
+
+    mapping(address => uint256) allocations;
     address payable public owner;
 
     /* constructor */
@@ -32,7 +33,7 @@ contract Fallout {
         msg.sender.transfer(address(this).balance);
     }
 
-    function allocatorBalance(address allocator) public view returns (uint) {
+    function allocatorBalance(address allocator) public view returns (uint256) {
         return allocations[allocator];
     }
 }
